+++ prototyper/CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file. See [conven
 - Add SpacemiT K1 SoC platform support for RustSBI Prototyper, including OrangePi RV2 board configuration.
 
 ### Modified
+- deps: update `sbi-spec` to version 0.0.10.
+- test-kernel: update PMU flag parameter trait names.
 - Refine CSR group comments.
 - fix(prototyper): temporary PMU fix for possible S-mode DTB modification
 - fix(prototyper): validate DBCN console shared memory range